Handle torrent data fields properly
The transmission RPC response is a well formed json document, all the
values have their actual contents, they are not pointers.

This commit allows the library to return proper values of torrent
attributes instead of the pointer addresses like in the past.

The code has also been simplified to avoid some intermediate marshalling
and unmarshalling steps.